+++ glmpowercalc/qprob.py
